@@ -42,60 +42,70 @@
     "tslib": "^2.0.0"
   },
   "devDependencies": {
+    "@aws-sdk/client-s3": "^3.496.0",
+    "@babel/core": "^7.14.5",
+    "@babel/preset-react": "^7.14.5",
     "@commitlint/cli": "^16.1.0",
     "@commitlint/config-conventional": "^16.0.0",
     "@mhlabs/cfn-diagram": "^1.1.33",
-    "@nrwl/cli": "13.10.2",
-    "@nrwl/cypress": "14.7.16",
-    "@nrwl/eslint-plugin-nx": "13.10.2",
-    "@nrwl/express": "13.10.2",
-    "@nrwl/jest": "13.10.2",
-    "@nrwl/linter": "13.10.2",
-    "@nrwl/node": "13.10.2",
-    "@nrwl/react": "13.10.2",
-    "@nrwl/web": "13.10.2",
-    "@nrwl/workspace": "13.10.2",
+    "@nx/esbuild": "17.2.8",
+    "@nx/eslint": "17.2.8",
+    "@nx/eslint-plugin": "17.2.8",
+    "@nx/jest": "17.2.8",
+    "@nx/js": "17.2.8",
+    "@nx/node": "17.2.8",
+    "@nx/react": "^17.2.8",
+    "@nx/webpack": "17.2.8",
+    "@nx/workspace": "17.2.8",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
     "@semantic-release/changelog": "^6.0.1",
     "@semantic-release/git": "^10.0.1",
     "@semantic-release/github": "^8.0.4",
-    "@testing-library/react": "13.4.0",
+    "@svgr/webpack": "^8.0.1",
+    "@swc-node/register": "~1.6.7",
+    "@swc/cli": "~0.1.62",
+    "@swc/core": "~1.3.85",
+    "@testing-library/react": "14.0.0",
     "@testing-library/react-hooks": "7.0.2",
     "@types/aws-lambda": "^8.10.93",
-    "@types/express": "4.17.13",
-    "@types/jest": "27.0.2",
-    "@types/node": "16.11.7",
-    "@types/react": "18.0.0",
-    "@types/react-dom": "18.0.0",
-    "@typescript-eslint/eslint-plugin": "~5.18.0",
-    "@typescript-eslint/parser": "~5.18.0",
-    "aws-cdk": "^2.43.1",
-    "aws-cdk-lib": "2.43.1",
+    "@types/express": "~4.17.13",
+    "@types/jest": "^29.4.0",
+    "@types/node": "18.16.9",
+    "@types/react": "18.2.33",
+    "@types/react-dom": "18.2.14",
+    "@typescript-eslint/eslint-plugin": "^6.9.1",
+    "@typescript-eslint/parser": "^6.9.1",
+    "aws-cdk": "^2.122.0",
+    "aws-cdk-lib": "2.122.0",
     "aws-cdk-lib-util": "^1.1.2",
     "aws-cdk-local": "^2.15.0",
-    "aws-sdk": "^2.1467.0",
-    "babel-jest": "27.2.3",
+    "babel-jest": "^29.4.1",
     "cdk-nag": "^2.12.32",
     "constructs": "^10.0.0",
-    "cypress": "^9.1.0",
     "cz-conventional-changelog": "^3.3.0",
     "dotenv": "^16.0.0",
-    "eslint": "~8.12.0",
-    "eslint-config-prettier": "8.5.0",
-    "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-import": "2.26.0",
-    "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-react": "7.29.4",
-    "eslint-plugin-react-hooks": "4.4.0",
-    "husky": "^7.0.4",
-    "jest": "27.2.3",
-    "nx": "13.10.2",
-    "prettier": "^2.5.1",
+    "esbuild": "^0.19.2",
+    "eslint": "~8.48.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-jsx-a11y": "6.7.1",
+    "eslint-plugin-react": "7.32.2",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "husky": "^8.0.3",
+    "jest": "^29.4.1",
+    "jest-environment-jsdom": "^29.4.1",
+    "jest-environment-node": "^29.4.1",
+    "nx": "17.2.8",
+    "prettier": "^2.6.2",
+    "react-refresh": "^0.10.0",
     "react-test-renderer": "18.0.0",
-    "semantic-release": "^19.0.5",
+    "semantic-release": "^23.0.0",
     "source-map-support": "^0.5.16",
-    "ts-jest": "27.0.5",
+    "ts-jest": "^29.1.0",
+    "ts-node": "10.9.1",
     "tsconfig-paths": "^3.14.1",
-    "typescript": "~4.6.2"
+    "typescript": "~5.2.2",
+    "url-loader": "^4.1.1"
   },
   "commitlint": {
     "extends": [
